Move `TenantEto` to the `Volo.Abp.MultiTenancy.Abstractions` project
