+++ framework/src/Volo.Abp.MultiTenancy.Abstractions/Volo.Abp.MultiTenancy.Abstractions.csproj
@@ -19,6 +19,7 @@
     <ItemGroup>
         <ProjectReference Include="..\Volo.Abp.VirtualFileSystem\Volo.Abp.VirtualFileSystem.csproj" />
         <ProjectReference Include="..\Volo.Abp.Localization\Volo.Abp.Localization.csproj" />
+        <ProjectReference Include="..\Volo.Abp.Auditing.Contracts\Volo.Abp.Auditing.Contracts.csproj" />
     </ItemGroup>
 
     <ItemGroup>
