+++ framework/src/Volo.Abp.MultiTenancy.Abstractions/Volo/Abp/MultiTenancy/TenantEto.cs
@@ -1,14 +1,16 @@
 ﻿using System;
 using Volo.Abp.Auditing;
+using Volo.Abp.EventBus;
 
-namespace Volo.Abp.TenantManagement;
+namespace Volo.Abp.MultiTenancy;
 
 [Serializable]
+[EventName("abp.multi_tenancy.tenant")]
 public class TenantEto : IHasEntityVersion
 {
     public Guid Id { get; set; }
 
-    public string Name { get; set; }
+    public string Name { get; set; } = default!;
 
     public int EntityVersion { get; set; }
 }
